moved and renamed managezoom to applyAestheticChanges
